thoth/requirements.txt
async-timeout==5.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
redis==5.2.0
requests==2.32.3
urllib3==2.2.3
websocket-client==1.8.0
